macaroons: add support for a Client ID (cid) caveat
Motivation:

Various use-cases involve a web-portal or other automated service
acquiring a macaroon and handing it (or a derivative) over to some
client.  The client will have authenticated to the web-portal (or
equivalent) but is not known to dCache.

It may be useful for the web-portal (or equivalent) to add some
identifier that is opaque to dCache, but has meaning to the web-portal.
This allows the portal to choose the granularity of the traceability;
e.g., the portal could add a unique ID as the 'cid' caveat value, or
could add the client's ID.  It also allows multiple agents to add 'cid'
values, provided care it taken to avoid namespace clashes.

Note that, as any client may add any caveat, any 'cid' caveat should not
be considered authorative without it being cryptographically bound to
the macaroon.  That said, if the web-portal ensures that it always adds
a cid caveat before handing over the macaroon, the first cid caveat
value is (fairly) reliable.

Modification:

Add support for extracting cid caveats and representing these as
ClientIdPrincipal objects in the logged in Subject.

Outside of the macaroon minting service, nothing in dCache makes use of
these principals.  Billing would be a natural place to record this
information.

Subsequent patches may add support for ClientIdPrincipal to include
other client-supplied identifiers: Globus transfer service, FTS and CMS
provide dCache with IDs that could include for logging purposes.

When authorising a macaroon-request with an existing macaroon, any
"macaroon" ClientIdPrincipals (i.e., 'cid' caveats) are copied into the
new macaroon, taking care to preserve their order.

Result:

Macaroons now support the 'cid' caveat.  This is ignored by dCache, but
should eventually allow for increased traceability.

Target: master
Require-notes: yes
Require-book: yes