webdav: add support for MaxUploadSize LoginAttribute
Motivation:

Add support for the MaxUploadSize LoginAttribute for
clients that use the WebDAV door.

Modification:

Discover any upload limitation placed on the authenticated user.

If the upload file size is supplied by the client, use this to fail
requests early, otherwise pass on the limitation to the pool so it can
enforce the limit for streaming uploads.

Result:

If a user has the MaxUploadSize LoginAttribute then this information is
used by the door to fail uploads if the file's size is known, otherwise
it is passed on to the pool for it to enforce the file size.

Target: master
Require-notes: yes
Require-notes: yes