ftp: add support for MaxUploadSize LoginAttribute
Motivation:

Add support in FTP doors for the MaxUploadSize attribute.

Modification:

Add code to accept MaxUploadSize attribute(s) returned as part of the
login and store the minimum value.

Add a pre-check when accepting data (the STOR, ESTO and PUT commands) so
that, if the client used the ALLO command to reserve capacity ahead of
an upload the transfer will fail fast if the file is too large.

Pass the maximum file size to the pool when initiating the transfer.

Minor adjustment to various error-handling code to support an upload
failing due to being too large.

Note there are two potential reply codes: 452 and 552.  The 452 value is
a temporary failure (e.g., filesystem is full).  The 552 value is a more
permanent failure (e.g., quota exceeded).

The 552 value is a better fit to the maximum upload size exceeded, and
is therefore used in the upload pre-check. The response from the pool is
insufficient to discover whether the problem is temporary or permanent
(e.g., the client retrying could select a different pool), therefore 452
is returned in this case.

Result:

Uploads that use FTP are subject to the user's MaxUploadSize
LoginAttribute(s), if any.

Target: master
Require-notes: yes
Require-book: yes