frontend: fix broken directory qos reporting
Motivation:

The frontend can report back to a client which QoS a file will recieve
when uploaded into a specific directory.  This calculation is wrong.

Modification:

Update the frontend to use the actual policy class that the pool
(likely) uses, when calculating the QoS of a directory.

Unfortunately, there is insufficient information to determine exactly
with which policy a file will be written; therefore, the output may
deviate from reality if a new file is accepted by a volatile pool (one
that has lsf mode precious).  That said, this patch provides a much
better description.

Result:

Frontend now more accurately describes the QoS of directories; i.e., the
QoS that newly written files will recieve when written into this
directory, assuming none of the targeted pools are volatile.

Target: master
Request: 4.2
Request: 4.1
Request: 4.0
Request: 3.2