frontend: fix reporting of files that are REPLICA ONLINE
Motivation:

Files written into a directory that has REPLCIA ONLINE will go to tape
with a copy stored on disk.  Currently, since there is no pin on this
file, we treat the file as being tape-only; however, the pool guarantees
the file is available with ONLINE latency, so we should report this as
disk+tape.

Modification:

Update frontend to use whether the file is pinned or should have system
sticky to determine the QoS.

Result:

Better reporting of the QoS of a REPLICA ONLINE file.

Target: master
Request: 4.2
Request: 4.1
Request: 4.0
Request: 3.2
Require-notes: yes
Require-book: no