xrootd: add support for anonymous fall-back for unknown users
Motivation:

SLAC xrootd server operates somewhat differently from dCache.  When a
client authenticates using GSI, it is sufficient for the client to
establish its identiy in terms of a DN from a certificate issued by a
trusted CA.  Subsequent operations (such as kXR_open) may be authorised
based on that identiy.

In contrast, dCache additionally requires the user is mapped to some uid
and one or more gids.  In general, this requires that all identities
must be known to dCache in advance: either as an individual or as some
group account.  Any attempt to use an unknown-to-dCache identiy will
fail, even though that identity is based on an EEC issued by a trusted
CA.

Perhaps the closes equivalent to the SLAC xrootd server behaviour in
dCache that is currently available is the anonymous user
(Subjects#NOBODY).  This would allow the user to list directories if
they are world readable, read data if the file is world readable, etc.

Currently, third-party copy operations from the receiving server are
authorised by a bearer token (the "rendezvous token"), not by that
server's identity.

If the receiving server uses an identity that is unknown to dCache then
dCache will reject the authentication.  However, this same server (using
the same credential) is able to receive data from SLAC xrootd servers.

This inconsistency points to a problem in dCache.

Modification:

Add UnionLoginStrategy to allow anonymous fall-back for GSI logins.

Result:

dCache xrootd door may be configured so it behaves more like the SLAC
xrootd server.

Target: master
Requires-notes: yes
Requires-book: yes