macaroon: add an issue ID caveat
Motivation:

Currently, a macaroon's ID is calculated from the macaroon's signature.
This has a couple of disadvantages:

First, all macaroons with the same set of caveats (and using the same
secret) will have the same ID and cannot be destinguished.

Second, if a client creates a new (derived) macaroon by autonomously
adding caveats to the dCache-issued macaroon then this derived macaroon
will have a different ID (which is good) but there is nothing linking
the new macaroon ID to the dCache-issued macaroon ID.

The first problem is somewhat mitigated by clients using a fixed
'validity' argument in the macaroon request, or by accepting dCache
built-in time limit.  This means that repeated requests will (likely)
have a different 'before' caveats, and so have distinct IDs.

The second problem becomes important when adding support for revoking a
specific macaroon.  Currently, storing the macaroon ID in a ban list is
not sufficient as any such revocation may be trivially by-passed by the
client adding a new caveat, as the new macroon would have a different
ID.

Modification:

dCache now add a 'iid' ("issue ID") caveat automatically.  The value is
a 48 bit random value that is base-64 encoded.

Update macaroon caveat parsing so that a valid macaroon can contain at
most one 'iid' caveat, which (if present) must be the first caveat.  The
ID of some macaroon (currently only used when logging) is now the 'iid'
caveat value, a colon, and the first 24-bits of the signature.

Backwards compatibility is maintained for already issued macaroons,
which do not have any iid caveat.  These macaroons will have the same ID
as before.

Result:

All dCache-issued macaroons now have a unique ID.  Any autonomously
derived macaroons can be associated with the origin (dCache-issued)
macaroon by considering the part of the ID before the colon.

Target: master
Requires-book: true
Requires-notes: true