WIP webdav/pool: add support for RFC3230 Want-Digest on COPY command
Motivation:

The COPY command can operate in PUSH or PULL mode.  In PUSH mode, the
data is already present in dCache while in PULL mode a file is created.

A client will likely care that data is transferred without being
corrupt.  Although dCache can (and by default, does) take steps to
ensure data integrety, the client may do this itself, too.  In
particular, FTS (in general) will always request the file's checksum.

Currently, a client can establish a file's checksum via RFC 3230
'Want-Digest' header on a GET or HEAD request, or by issuing a PROPFIND
to discover the checksum value.  This means that, after the transfer,
the client must issue a GET or (more likely) a HEAD request to obtain
the checksum.

There are two problems:

  1. the pool may not have calculated the client's desired checksum.

  2. the HEAD/PROPFIND request is an extra request, which means
     delays and the risk of a macaroon expiring.

Modification:

Add support for RFC 3230 'Want-Digest' on COPY requests.

For PUSH requests, this is trivial.  Support is similar to that for GET
and HEAD requests.

For PULL request, adding support has two parts.  First, the desired
checksum is relayed to the pool, guaranteeing that (provided dCache
supports the algorithm) that checksum value will be available.  Second,
the output is updated to support the 'Digest' response as an HTTP
trailer.  HTTP trailers are HTTP Headers that appear after the entity,
rather than before.  This is necessary as the checksum is only known
after the transfer is complete.

Result:

A client can issue a COPY request (PUSH or PULL) without requiring
an additional request to discover the file's checksum.

It also allows the client to hint which checksum algorithm dCache should
support for future RFC 3230 HEAD or GET requests.

Target: master
Requires-notes: no
Requires-book: no