pool: HTTP TPC create 'Want-Digest' value based on known checksums
Motivation:

When verifying an HTTP PUT request, dCache issues an HTTP HEAD request
to fetch checksums of the freshly created file.  It currently asks for a
fixed set of checksums: ADLER32 and MD5.

There is a risk that the remote site replies with an MD5 checksum when
dCache doesn't have an MD5 checksum value for this file.

Modification:

Update the HEAD request requests so it only requests checksum values
that dCache already knows.

Result:

Avoid requesting a checksum value from the remote server that dCache
cannot then use to verify the integrety of the transfer.

Target: master
Requires-notes: no
Requires-book: no