pool: add inotify support for open/read/write/close events
Motivation:

The inotify API supports clients receiving events for when a file is
opened, when data is read from or written to an open file, and two types
of close event.

dCache currently sends none of these events.  There may be use-cases
where clients may depend on this information.

Modification:

The pool is the main focus of this patch, adding the ability for pools
to generate inotify events.

The inotify API allows clients to monitor all items within a directory.
Therefore, pools need to know the parent directory of the targeted file.

PnfsManager now adds the links (parent directory and target name) as the
'links' element of the StorageInfo map.  The pool ensures this
information is not written to the repository.

A mechanism is introduced that allows a ReplicaRecord to be aware of the
PnfsManager-generated FileAttributes, even when opening an existing
replica.

This patch introduces InotifyChannel, InotifyReplicaRecord,
InotifyReplicaStore to instrument repository channel, to be aware of
client activity and emit the appropriate IN_OPEN, IN_ACCESS, IN_MODIFY,
IN_CLOSE_WRITE and IN_CLOSE_NOWRITE events.

Care is taken to avoid triggering events on any dCache-internal
activity, such as checksum calculation, pool-to-pool transfers and
nearline flush or stage operations.

Result:

inotify clients can now receive events descring opening a file, writing
data or reading the data, and closing the file: IN_OPEN, IN_ACCESS,
IN_MODIFY, IN_CLOSE_WRITE, IN_CLOSE_NOWRITE.

Target: master
Requires-notes: yes
Requires-book: no