cells: add support classes to learn in which zone they reside
Motivation:

Some cells need to know in which zone they reside in order to propagate
that information or adjust their behaviour.  A cell using
UniversalSpringCell does not have direct access to CellNucleus,
therefore must discover this information by implementing an interface.

Modification:

Add interface to discover in which zone a cell resides.

Update UniversalSpringCell to inject this value if a cell requires it.

Result:

No admin or user observable change, but dCache now supports cells
learning in which zone they reside when the cell uses the
UniversalSpringCell.

Target: master
Requires-notes: no
Requires-book: no