pool: add 'zone' tag
Motivation:

A pool publishes tags that describe itself.  One interesting aspect of a
pool is its geographical location, as represented by a zone.

Modification:

Introduce the concept of reserved tags: tags that the admin may not
specify.

Include the zone within which the pool resides as the tag 'zone'.

Result:

A pool now includes a 'zone' tag where the value is zone within which
the pool resides.

Target: master
Requires-notes: no
Requires-book: no