info: publish a domain's zone
Motivation:

The info service collects information about all dCache components.  This
should include a domain's zone.

Modification:

Update the static information publisher to silently skip metric with no
data.

Publish a domain's zone in the static information block, as the 'zone'
metric.

Result:

The info service now list a domain's zone as the metric
'domains.<domain-name>.static.zone', if the domain is configured to have
a zone.  If the domain is not configured with a zone then this metric is
not published.

Target: master
Requires-notes: yes
Requires-book: yes