poolmanager: support querying a pool's zone
Motivation:

Various tasks may require knowledge of in which zone a pool resides. To
support this, poolmanager must remember this information and provide it
on demand.

Modification:

Update PoolMonitor interface to suppy zone information for a file.

Update a pool's representation in poolmanager so that it contains a
pool's zone.

Result:

No user or admin observable changes, but PoolManager knows in which zone
a pool resides, and allows other code to query this information.

Target: master
Requires-notes: no
Requires-book: no