pool: add alternative load monitoring
Motivation:

Since we want to remove queues, we need some way of measuring how busy
is a pool that does not depend on the concept of queues.

Modification:

Introduce LoadObserver interface, which provides an abstract way of
obtaining the current load on a pool.

Provide a LoadObserver implementation that monitors threads as they make
IO requests.  It keeps track of the number of in-flight IO threads over
time and calculates the average number over some history period.  This
is compared to some maximum number of in-flight IO requests (with which
the storage system is expected to cope) to calculate the load.

Result:

Pools now have the ability to measure their IO load.

Target: master
Requires-notes: yes
Requires-book: yes