docs: TheBook rewrite the install chapter
Motivation:

The install chapter is perhaps the first chapter perspective new dCache
admins are likely to see: in essence it is our chance to "sell" dCache.
The chapter must make it be easy to install, and should show off (or
hint at) some of the advanced features of dCache.

There are many things wrong with the current install chapter:

   * it contains wrong information,

   * cannot work as steps are in the wrong order,

   * contains too few checks to verify progress so far,

   * poorly structured: little sense of order or progression

   * heavily X.509 centric, which both promotes the wrong image
     and makes dCache harder to deploy,

   * stops at a single-node deployment, with no progress beyond.

Fixing these problems is more than a few changes here and there: a
sestemic approach is needed.  Therefore, this commit completely rewrites
the install chapter.

The new install chapter is split into distinct steps, with a goal for
each step.  The authentication is focused on username+password as this
has the lowest bar-to-entry, but includes information on how to
configure X.509 and OpenID-Connect authentication.  Covers host
certificates, but doesn't require them to have a working dCache
instance.  Documents how to obtain host certificate without requiring
access to an IGTF CA.

Currently missing:

  * There is currently nothing on how to configure VOMS-based
    authentication.

  * The sections on multi-host and high-available dCache are largely
    place-holders.

  * Adding GridFTP support is missing.

  * Many of the worked examples are missing: FTPS, GridFTP, X.509 client
    authentication

  * SRM may be added