common: add support for interface reachability configuration
Motivation:

The FTP protocol requires the server to tell the client to which
interface it can connect.  A client connected to the control channel
with one address family may request data be transferred with another
family; for example, a client with a control channel connected to the
server using an IPv6 address requests data be transferred using IPv4.

The correct interface cannot be discovered in a fully automatic fashion
without making assumptions.

Currently dCache assumes that the address to which the client connects
and the desired data channel are on the same interface.  If no suitable
address is available on the same interface then dCache will fail the
request.

Although this works with many deployments, this has proved too
restrictive in some cases, where IPv4 and IPv6 addresses are places on
different interfaces.

Modification:

Add support for explicit configuration of which interface corresponds to
which subnet.  This is achieved through system properties with the form:

    org.dcache.net.interface-reachability.<n>.subnet = <CIDR>
    org.dcache.net.interface-reachability.<n>.interface = <address>

Where <n> is an integer, <CIDR> is the subnet from which the client is
connecting and <address> is the network interface address.  Both the
<CIDR> and the <address> values may be (independently) IPv4 or IPv6
addresses.  The values of <n> must be consecutive, starting with 1.

Here is a complete example:

    -Dorg.dcache.net.interface-reachability.1.subnet=198.51.100.0/24
    -Dorg.dcache.net.interface-reachability.1.interface=2001:db8::206
    -Dorg.dcache.net.interface-reachability.2.subnet=203.0.113.0/24
    -Dorg.dcache.net.interface-reachability.2.interface=2001:db8::208
    -Dorg.dcache.net.interface-reachability.3.subnet=2001:db8::/32
    -Dorg.dcache.net.interface-reachability.3.interface=192.0.2.25

Result:

FTP is supported in dual-home IPv4/IPv6 (dual stack) deployments.  This
patch is based on a patch originally by Francesco Prelz.  We thank him
for his contribution.

Target: master
Requires-notes: yes
Requires-book: yes
Request: 6.0
Request: 5.2
Request: 5.1
Request: 5.0
Request: 4.2