pool: httptpc support debugging transfers
Motivation:

HTTP-TPC transfers are occationally problematic.  This may require some
detailed diagnostic information to figure out what went wrong.

Some times the problem is with the interactions with the remote server.
However, network connections are encrypted to prevent evesdropping.
Therefore the HTTP request/response logging must take place within
dCache (or the remote party).

Other problems could be related to poor performance.  Resolving such
issues may require timing information and information about how well the
filesystem performed during the transfer.

This is considerable additional information that cannot be provided for
all transfers.  To be useful in production environments, the set of
transfers (for which additional information is provided) must be
selective.  dCache must allow the admin to choose which transfers are
under investigation.  The remaining transfers continue normally, without
any additional logging.

Modification:

Instrument the pool's HTTP-TPC implementation to provide additional
information.

The HTTP-TPC mover is updated to optionally log the HTTP-TPC requests
made by dCache along with the corresponding responses.  In addition, the
filesystem IO statistics are also optionally logged.

Add ability to select transfers for which additional logging is enabled.

Transfers are identified as something where debug information should be
logged through a set of rules.  Each rule provides match criteria
against the dCache absolute path ("local-path"), the remote server's
host name (as specified by the HTTP-TPC initiating client), and the
remote server absolute path ("remote-path").

Rules may be specified on the pool (involved in the transfer) or in the
WebDAV door (to which the client connected).  The pool will log debug
information if a rule matches on either the door or the pool.

The patch includes a common set of admin commands (pool and WebDAV door)
that allow the amdin to manage the set of rules.

Result:

It is now possible for an admin to select HTTP-TPC transfer (on either
the WebDAV door or on the pool) for which the pool will log additional
diagnostic information.

Target: master
Requires-notes: yes
Requires-book: yes