Add support for inspection
Motivation:

There are situations where we want dCache to provide more detailed
logging information; these are situations where something has gone wrong
or where dCache is not working quite right.

This concept already exists in dCache, albeit as ad-hoc examples.  One
such example is the enhanced logging provided when a GridFTP transfer is
aborted.  Another example is the extended logging provided when a login
fails.

We would like a framework within dCache that decouples the decision to
provide enhanced logging from the code that provides that information.
This is to allow more sophisticated logging possibilities; for example,
provide detailed information about the next HTTP transfer from an
specific IP address, targeting a particular file.

The information should be bound to the CDC since this context normally
matches an activity that should be inspected.

Modification:

Add a machanism to CDC where code can state it wants to enable dCache
inspection, and allow code to discover if inspection is enabled.

The code requires that a short explanation is specified when inspection
is enabled.  This is to allow a more meaningful message when logging the
additoinal information.

Result:

No user or admin observable changes.

Target: master
Requires-notes: no
Requires-book: no