pool: add ability to react to the inspect flag at end of a transfer
Motivation:

When a transfer has ended (abnormally or otherwise) the admin may wish
to know some details about that transfer.

Modification:

Add an "inspection point" at the end of a transfer, where details of a
transfer are recorded.

Currently no movers are instrumented to provide this information, so
this has no effect; additionally, there's currently no way to switch on
the inspection flag.

Result:

Add "inspection point" for logging details of a transfer when it has
finished (successfully or otherwise).

No user or admin visible changes.

Target: master
Requires-notes: no
Requires-book: no