pool: ftp enable inspection if client aborted transfer
Motivation:

If the user aborted a transfer then we want to automatically enable
inspection, so that the details may be logged.

Modification:

Add code to enable inspection so that a transfer is inspected once it is
completed.

Note that this currently does nothing, but subsequent changes will
switch the current hard-coded mover status logging to use Discribable
framework.

Result:

Inching closer to something useable, but still no user or admin
obserable changes

Target: master
Requires-notes: no
Requires-book: no