poolmanger: add unit test for RequestContainerV5
Motivation:

RequestContainerv5 currently has no unit-tests.  This makes refactoring
potentially dangerous.

Modification:

Add a set of unit tests.  The primary focus of these tests is the
finite-state-machine and the current code achieves as close to 100% code
coverage as possible.  There are several code paths that are impossible
to exercise without simulating non-existing bugs within
RequestContainerV5 (i.e., dead-code).

Each unit test is written using the BDD style, with clearly demarked
"given" (the setup or fixture), "when" (the trigger being tested) and
"then" (assertion) sections.

The exception are tests that trigger a request being suspended.
RequestContainerV5 provides no convenient way to determine that a
request is suspended (beyond notifying billing).  For these tests, the
admin command to retry a request is used to trigger a response to the
door.  Therefore, for these tests there are two "when" sections.

The support for checking whether a pool is still processing a
pool-to-pool or stage request ("ping") within RequestContainerV5 is also
updated, to support a one-shot mode.  This is used by the unit-testing
so that, under test conditions, RequestContainerV5 sends the "ping"
request exactly once, reliably.

Result:

No user- or admin observable changes.

Code is more robust against regressions from code refactoring.

Target: master
Requires-notes: no
Requires-book: no