gplazma: oidc check token validity if JWT
Motivation:

Many OIDC tokens are formatted using JWT.  Although the OIDC plugin
doesn't (yet) support offline verification, a JWT token may parsed and
information extracted.  Although the veracity of the token isn't known,
some of the claims may still be used.

The plugin already uses the 'iss' claim to identify which OP issued the
token.  This allows the plugin to send the token to a single OP's
user-info endpoint.

The not-before ('nbf') and expiry ('exp') claims may also be used
without verifying the token: if a malicious user presents a modified
token (e.g., an expired token with modified 'exp' claim), dCache will
call the user-info endpoint and discover the problem.

By checking the nbf and exp claims explicitly (in the plugin) then we
get better behaviour if a token has expired (or is not yet valid): there
is a faster response (no network interactions), less load on the OP, and
better error messages.

Modification:

If the oidc token is a JWT then check the token hasn't expired before
attempting to contact the OP's userinfo endpoint.

Result:

The oidc plugin has been updated to provide better handling of expired
access tokens that are JWT.

Target: master
Request: 7.0
Requires-notes: yes
Requires-book: no