core: fix macaroon behaviour with root and path
Motivation:

A macaroon may contain a multiple of both "root:" and "path:" caveats.
These interact in a somewhat complex fashion.  The MacaroonContext class
takes care to process these caveats correctly.

An important aspect is that the overall resulting (or "effective")
"path" path must be interpreted as being relative to the result (or
"effective") "root" path.

The MacaroonLoginStrategy mistakenly interpreted the "path" path as an
absolute path in dCache when adding the PrefixRestriction, since this
accepts an absoluate path in dCache namespace.

This mistake is somewhat explained by MacaroonContext#getPath mistakenly
returning FsPath, which is inteded only for absolute paths in dCache's
namespace.

If there is no "root:" caveat then there is no difference and any
"path:" caveat will work as expected.  Similarly, a macaroon without a
"path:" caveat but with a "root:" caveat will also work as expected.

However, if a macaroon has both a "root:" and "path:" caveats then the
macaroon does not work as expected.

Modification:

Add JavaDoc and update return type of MacaroonContext to make clear the
semantics of the returned value.

Update unit-tests accordingly.

Update MacaroonLoginStrategy to resolve the effective "path" path
against the effective "root" path.

Result:

Macaroons created with both a "path" caveat and "root" caveat work as
expected.

Target: master
Request: 7.2
Request: 7.1
Request: 7.0
Request: 6.2
Request: 6.1
Request: 6.0
Request: 5.2