common: have ParsableFile return Result
Motivation:

Currently, ParsableFile#get returns the parsed file's contents wrapped
within an Optional.  The Optional is needed because reading the file
might fail; for example, the file might be missing or have invalid
contents (from which the parser is unable to recover).

Using Optional means that the code using ParsableFile can learn when
there was a problem reading the file, but not what was that problem.

There is existing code in dCache that reads files and that is not
currently using ParsableFile.  That code must propagate a description of
any parse failures.  This is not possible with Optional.

Modification:

Introduce a new class Result.  This is broadly similar to Optional but
it carries information about what went wrong if the operation fails. The
design of Result was strongly influenced by similar classes in other
languages (e.g., Either in Scala and Haskell, Result in elm and rust)
and is a simple example of the ADT "sum"-type.

Note that, although there is some overlap with {@link
java.util.concurrent.CompletionStage} (for example, compare
Result#consume with CompletionStage#whenComplete), Result is much
simpler (it does not support asynchronous operations and does not
support composition) and more flexible as it doesn't require that
failures are represented by a Throwable.

The existing code is updated to use Result instead of Optional.

Result:

No user- or admin observable changes.  The ParsableFile code is updated
in preparation for future adoption.

Target: master
Require-notes: no
Require-book: no