gplazma: oidc allow suppression of offline verification
Motivation:

Some OPs (e.g., EGI-CheckIn) provide more information about the user via
the userinfo endpoint that is encoded in the JWT.  If that additional
information is important then dCache must use the userinfo endpoint.

Modification:

When attempting to verify a JWT, check whether the OP is configured to
suppress offline verification.  If so, fail the offline verification and
fall back to querying the userinfo endpoint.

Result:

An OP declaration now supports the `-suppress=offline` argument to force
dCache to query the userinfo endpoint, even if the access token is a JWT.

Target: master
Requires-notes: yes
Requires-book: no
Request: 8.2