chimera: add methods to determine whether an FsInode is the root
Motivation:

In forthcoming patches, it will prove useful to know whether a given
FsInode represents the root directory of the filesystem.

Modification:

An unusued method `isRoot` is removed from `FsInode_SURI`.

Update the FileSystemProvide interface to expose a method for testing
whether some given FsInode object is the root directory.  This method is
implemented for JdbcFs.

A new method is added to FsInode to expose this information.

Some unit tests are added.

Result:

No user or admin observable changes.  An FsInode object may now be
queried to determine whether it is the root directory.