Chimera: include name of inode under parent.
Motivation:

The FsInode class represents the inode of a file or directory.  It
already optionally records the parent directory's inode: for
directories, this inode is unambigous; other items, the recorded parent
could be only one of several parents (due to hard links).

In a subsequent patch, it proves useful to know the name of the FsInode
under the recorded parent directory.

Modification:

Update FsInode to support querying the name under which it is stored.

With this patch, the method that allows an FsInode object to learn its
parent directory is updated to accept the name, too.  The places where
this method is called are updated to provide this information.

A minor refactoring is introduced so that when querying an FsInode to
learn either the parent FsInode or this FsInode's name (under that
parent) and this FsInode does not already have this information then a
database look up is triggered.  This is consistent with existing
behaviour.

A minor optimisation is introduced: if the directory is the root
directory then no database lookup is made when querying either the
parent inode or this inode's name.

Result:

No user- or admin observable changes.  An FsInode object may now be
queried to discover its name within a parent directory.

Target: master
Requires-notes: no
Requires-book: no