gplazma: update ExemptFromNamespaceCheck to include allowed prefixes
Motivation:

The ExemptFromNamespaceCheck principal indicates that this login session
should be exempt from the normal namespace permissions model; instead,
the user is given Carte Blanche but only for specific authorisations,
where an authorisation is a subtree (a path prefix) and a set of
operations.

This limitation is normally encoded as a Restriction; however, for
rather prosaic reasons, it is helpful if the ExemptFromNamespaceCheck
itself contains the path prefixes.

Modification:

Update ExemptFromNamespaceCheck to support carry the prefixes.  The
principal now also supports querying whether a path is allowed.

Update the oidc plugin to include the allowed paths (from the set of
authorisations) when adding an ExemptFromNamespaceCheck principal.

Update Gplazma2LoginStrategy to support adding the upload path to
ExemptFromNamespaceCheck.

Result:

No user or admin observable changes.  The ExemptFromNamespaceCheck
principal (if present) allows for testing whether it is allowed to
access a particular path.

Target: master
Requires-notes: no
Requires-book: no