webdav: limit requested fileattributes to those needed
Motivation:

Currently, when the WebDAV door fetches information from PnfsManager to
support a PROPFIND request, it does so ignoring what information the
client has actually requested.  This results in (quite likely) an overly
broad selection of information, which makes the PnfsManager request
slower than it needs to be.

Modification:

Milton does not provide the list requested properties when making the
getChildren request.  Therefore we must introduce a work-around to allow
dCache to discover this information; in this case, it is a thread-local
variable.  A Milton class is wrapped to ensure the information is
up-to-date.

Further complicating this is that Milton checks the existence of the
PROPFIND target by requesting the corresponding Resource.  It does this
ahead of identifying the desired properties.  This patch leaves the
current behaviour (request sufficient information for all possible
properties); a subsequent patch can improve upon this.

Added a table that maps WebDAV properties to a corresponding set of
FileAttribute (needed to support that property).  This is used to build
the minimum set of FileAttributes needed to support the request.

The XATTR FileAttribute is moved from the essential selection to those
needed for GET and HEAD requests.

Result:

PROPFIND requests with Depth 1 (or larger) are now faster.

Target: master
Requires-notes: no
Requires-book: no