chimera: add extra unit-tests for JdbcFs
Motivation:

A previous patch (2fb9170283) introduced a bug that resulted in data
being lost when uploaded using xrootd with the kXR_POSC flag and that
overwrites an existing file.  This problem was because the patch
introduced a bug in the \`JdbcFs#rename\` method.

This specific problem identified a more general problem with a lack of
unit-tests for JdbcFs.

Modification:

Introduce extra unit-tests to cover the methods that patch 2fb9170283
touched.

Results:

No user- or admin observable changes.

The behaviour of JdbcFs is now more robust against problems introduced
by refactoring.

Target: master
Requires-notes: no
Requires-book: no