gplazma: oidc more descriptive offline verification failures
Motivation:

OIDC uses caches to record the OIDC discover document and the issuer's
public keys (for offline varification).

Currently, if there is a problem with the discovery document or with the
JWKS document then those problems are logged when the document is
loaded, but subsequent attempts for offline verification of a token fail
with an opaque and cryptic message.

Modification:

Update Result to support additional (Option-like) methods: `map` and
`flatMap`.

Introduce a new MissingNode class (ReasonBearingMissingNode) that can
carry the reason why a JSON node is missing.  Update discovery document
fetching to take advantage of this new class.

Update MemoriseMapWithExpiry to work on Results rather than Maps.  The
class is renamed to reflect this change.

Issuer is updated to work with Result objects rather than Optional
objects.  These Result objects, if representing a failure, describe why
the operation failed.  This is used to generate a more descriptive
AuthenticationException message.

Suppress hamcrest-core from junit as this brings in an ancient
version.

Result:

During offline verification, the details regarding any failure to fetch
documents from the issuer or to parse the issuer-supplied documents are
now used to build the error message when the plugin rejects a token.

Target: master
Requires-notes: yes
Requires-book: no
Closes: #7553