gplazma: update explain login to allow credentials
Motivation:

A common question is to understand how a bearer token (such as an OIDC
access token) is processed by gPlazma.  Given that a bearer token may be
simply copied into the admin shell, the `explain login` command may be
extended to support this query.

This patch is based on an idea from Onno Zweers.

Modification:

Update `Subjects.subjectFromArgs` to support arguments that describe
credentials, with a single type of credential (bearer token) given by
the prefix `token`.

Add corresponding unit tests

Update `explain login` documentation to describe how to see the result
of a login.

Result:

The admin is now able to query gPlazma with a bearer token and discover
the result of that login process, without involving any doors.

Closes: #7576
Target: master
Requires-notes: no
Requires-book: perhaps